% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_grid_data.R
\docType{data}
\name{elevation_grid}
\alias{elevation_grid}
\title{An R dataframe of latitude and longitude over the continental United States.}
\format{
A data frame with 3 columns and 1372 rows

\describe{
 \item{x}{a numeric vector of longitude values}
 \item{y}{a numeric vector of latitude values}
 \item{elevation}{a numeric vector of elevation values}
 }
}
\usage{
elevation_grid
}
\description{
The dataframe of points is composed for a resolution 50 grid.
}
\keyword{datasets}
