% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_interpolation.R
\name{graph_interp}
\alias{graph_interp}
\title{A function that graphs an interpolation of values over the US.}
\usage{
graph_interp(preds, grid)
}
\arguments{
\item{preds}{A vector of predictions to be interpolated over the US grid.
Preds should be generated from the `interpolate_grid` function.}

\item{grid}{A dataframe with columns x and y that represents the US grid.}
}
\value{
A ggplot2 object that graphs the interpolated values over the US
grid.
}
\description{
The function takes a vector of predictions and graphs them over the US grid
at a given resolution. The function uses the `grid_usa` function from seesaw
to plot the predictions over the US grid.
}
\examples{
grid <- usagrid(50)
preds <- runif(nrow(grid),0,20)
graph_interp(preds, usagrid(50))

}
