% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_info_data.R
\docType{data}
\name{station_info}
\alias{station_info}
\title{NOAA Weather Station Attributes in the U.S.}
\format{
A data frame with 5 columns and 237 rows
\describe{
 \item{station_id}{The unique identifier for each weather station}
 \item{station_name}{The name of the weather station}
 \item{state}{The state in which the weather station is located}
 \item{longitude}{The longitude of the weather station}
 \item{latitude}{The latitude of the weather station}
 }
}
\usage{
station_info
}
\description{
A dataframe containing the attributes of NOAA weather stations in the U.S.
including lat, lon, state, station name, and station id.
}
\keyword{datasets}
