% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series_station}
\alias{time_series_station}
\title{Extract the time series for a specific station by station id.}
\usage{
time_series_station(id, start_date = 11275, end_date = 19820)
}
\arguments{
\item{id}{The station id to extract the time series for.}

\item{start_date}{The start date to extract the time series for. Expects a
date in the format of the LST_DATE column in the full_table data frame. This
format is "YYYY-MM-DD".}

\item{end_date}{The end date to extract the time series for. Expects a date
in the format of the LST_DATE column in the full_table data frame. This
format is "YYYY-MM-DD".}
}
\value{
A data frame containing the time series for the specified station.
}
\description{
Obtain the NOAA weather forecast data for a specific station between the
years 2000 and 2024.
}
\examples{
Durham, NC
time_series_station(54794, "2000-01-01", "2010-12-31")

}
