% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_X_data.R
\docType{data}
\name{training_X}
\alias{training_X}
\title{An R dataframe of latitude and longitude over NOAA weather stations.}
\format{
A data frame with 4 columns and 135 rows.

\describe{
 \item{(Intercept)}{a numeric vector of intercept values}
 \item{x}{a numeric vector of longitude values}
 \item{y}{a numeric vector of latitude values}
 \item{elevation}{a numeric vector of elevation values}
 }
}
\usage{
training_X
}
\description{
The dataframe of points is composed for NOAA stations that existed between
2000 and 2024.
}
\keyword{datasets}
