% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_trends.R
\name{trend_of_temps}
\alias{trend_of_temps}
\title{A function for estimating the trend of temperatures over time,
in units of degrees Celsius per year.}
\usage{
trend_of_temps(station_id, date_start = "2000-01-01", date_end = "2024-12-31")
}
\arguments{
\item{station_id}{The WBANNO of the station to estimate the temperature trend for, formatted
as a character. Expects a WBANNO present in the station_info data frame.}

\item{date_start}{A character vector of the start date for the analysis. It
should be in the format "YYYY-MM-DD".}

\item{date_end}{A character vector of the end date for the analysis. It should
be in the format "YYYY-MM-DD".}
}
\value{
A named list with the trend (defined as the average increase in
temperatures on the first of the month over the dates specified) and standard
error. Will return a list with the trend and standard error as NA if the
station is less than two years old.
}
\description{
A function for estimating the trend of temperatures at one station over a
specified period of time, given in units of degrees Celsius per year. Function
also returns a standard error for the estimate.
}
\examples{
trend_of_temps("53878", "2000-01-01", "2020-12-31")

}
