% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_usa.R
\name{usagrid}
\alias{usagrid}
\title{Create a grid of points over the contiguous United States}
\usage{
usagrid(resolution = 50)
}
\arguments{
\item{resolution}{The number of points to generate in each dimension of the
grid.}
}
\value{
A data frame containing the grid points within the contiguous
United States.
}
\description{
This function creates a grid of points over the contiguous United States.
The grid points are generated within the bounding box of the contiguous
United States and does not include points over Alaska, Hawaii, or any of
the territories.
}
\examples{
grid_usa(100)

}
