% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_cycle.R
\name{yearly_cycle_station}
\alias{yearly_cycle_station}
\title{A function for estimating the yearly cycle for one station.}
\usage{
yearly_cycle_station(id, variable = "T_DAILY_AVG")
}
\arguments{
\item{id}{The WBANNO of the station to estimate the yearly cycle for,
formatted as a character. Expects a WBANNO present in the station_info data
frame.}

\item{variable}{The variable to estimate the yearly cycle for.
Default is "T_DAILY_AVG". Expects a variable present in the full_table
data frame.}
}
\value{
A data frame with the expected variable value for each day of the
year.
}
\description{
A function for estimating the yearly cycle for one station. A yearly cycle is
simply the expected temperature on each day of the year.
}
\examples{
yearly_cycle_station("03047")

}
