% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded_interpolation.R
\name{interpolate_grid}
\alias{interpolate_grid}
\title{A function for interpolating data from the stations to a grid points within
the contiguous USA.}
\usage{
interpolate_grid(X, y, Xpred = NULL, resolution = 50)
}
\arguments{
\item{X}{The design matrix. Expects a dataframe with one column of 1s for
the intercept, an 'x' column for the longitude, and a 'y' column for the
latitude. Additional columns can be included to account for other
covariates.}

\item{y}{The response variable that would like to be interpolated. Must be
a numeric vector with the same number of rows as X.}

\item{Xpred}{The design matrix for the prediction grid. If NULL, the function
will create a grid of points over the contiguous USA at the
specified resolution. Xpred expects a dataframe with one column of 1s for
the intercept, an 'x' column for the longitude, and a 'y' column for the
latitude. Additional columns can be included to account for other covariates.}

\item{resolution}{The resolution of the grid to interpolate the variable over.}
}
\value{
A data frame containing the interpolated values for the specified
variable over the contiguous USA.
}
\description{
The function takes in a design matrix, response variable, resolution,
and optional design matrix for the prediction grid (if additional covariates
are desired) as input to interpolate that variable over the lower 48 states.
}
\examples{

X <- data.frame(x = c(runif(100,min = -102, max = -81)), y = c(runif(100,min = 31, max = 37)))
X <- as.data.frame(cbind(1, X))
y <- c(runif(100, min = 0, max = 20))
interpolate_grid(X, y, Xpred = NULL, resolution = 20)

}
