% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded_interpolation.R
\name{interpolate_grid}
\alias{interpolate_grid}
\title{A function for interpolating data from the stations to a grid points within
the contiguous USA.}
\usage{
interpolate_grid(start_date, end_date, var = "T_DAILY_AVG", resolution = 50)
}
\arguments{
\item{start_date}{The start date to interpolate the variable for. Expects a
date in the format of the LST_DATE column in the full_table data frame. This
format is "YYYY-MM-DD".}

\item{end_date}{The end date to interpolate the variable for. Expects a date
in the format of the LST_DATE column in the full_table data frame. This
format is "YYYY-MM-DD".}

\item{var}{The variable to interpolate. Default is "T_DAILY_AVG". Variable
must be a column in the full_table data frame. For a list of variables, see
the full_table data frame documentation.}

\item{resolution}{The resolution of the grid to interpolate the variable over.}
}
\value{
A data frame containing the interpolated values for the specified
variable over the contiguous USA.
}
\description{
The function takes a start, end date, variable, and resolution as input to
interpolate that variable over the lower 48 states during that time period.
}
\examples{
interpolate_grid("2024-03-01", "2024-03-31", "T_DAILY_AVG", 50)

}
