% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_trends.R
\name{trend_of_temps}
\alias{trend_of_temps}
\title{A function for estimating the trend of temperatures over time,
in units of degrees Celsius per year.}
\usage{
trend_of_temps(station_id, date_start = "2000-01-01", date_end = "2024-12-31")
}
\arguments{
\item{station_id}{The WBANNO of the station to estimate the temperature trend for, formatted
as a character. Expects a WBANNO present in the station_info data frame.}

\item{date_start}{A character vector of the start date for the analysis. It
should be in the format "YYYY-MM-DD".}

\item{date_end}{A character vector of the end date for the analysis. It should
be in the format "YYYY-MM-DD".}
}
\value{
A data frame with the station ID and a column for each month of the year.
The column value represents the slope of the temperature trend in degrees Celsius
of the first day of the month over the time period specified. One final column
in the dataframe stores the mean of the monthly slopes.
}
\description{
A function for estimating the trend of temperatures over time,
in units of degrees Celsius per year.
}
\examples{
trend_of_temps(c("53878","04130"), "2000-01-01", "2020-12-31")

}
